Change some usage of RawKeyEvent to KeyEvent in preparation for deprecation
