+++ examples/api/lib/widgets/focus_manager/focus_node.0.dart
@@ -41,7 +41,7 @@ class _ColorfulButtonState extends State<ColorfulButton> {
     super.initState();
     _node = FocusNode(debugLabel: 'Button');
     _node.addListener(_handleFocusChange);
-    _nodeAttachment = _node.attach(context, onKey: _handleKeyPress);
+    _nodeAttachment = _node.attach(context, onKeyEvent: _handleKeyPress);
   }
 
   void _handleFocusChange() {
@@ -52,7 +52,7 @@ class _ColorfulButtonState extends State<ColorfulButton> {
     }
   }
 
-  KeyEventResult _handleKeyPress(FocusNode node, RawKeyEvent event) {
+  KeyEventResult _handleKeyPress(FocusNode node, KeyEvent event) {
     if (event is RawKeyDownEvent) {
       debugPrint('Focus node ${node.debugLabel} got key event: ${event.logicalKey}');
       if (event.logicalKey == LogicalKeyboardKey.keyR) {
